% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_graph.R
\name{ci_graph}
\alias{ci_graph}
\title{Create cumulative incidence plots}
\usage{
ci_graph(
  model_name,
  outcome_names = NULL,
  outcome_colors = NULL,
  col_palette = "Dark 3",
  xlab_name = "Time",
  ylab_name = "Cumlative Incidence",
  conf_int = TRUE,
  conf_int_alpha = 0.2
)
}
\arguments{
\item{model_name}{A competing risks object of the class survfitms survfit}

\item{outcome_names}{A vector of the outcome names, default is 1, 2, ... n}

\item{outcome_colors}{A vector consisting of the color names/hexadecimals. Default is colors from Dark 3 palette}

\item{col_palette}{A string containing the name of the hcl palette to use. See hcl.pals() for names. Default is Dark 3}

\item{xlab_name}{Label for the x-axis. Default is "Time"}

\item{ylab_name}{Label for the y-axis. Default is "Cumulative Incidence"}

\item{conf_int}{Determines whether a confidence interval should be drawn for each curve. Default is TRUE.}

\item{conf_int_alpha}{A value between 0 and 1 that determines the alpha of the ribbon confidence interval}
}
\value{
A GGPlot object with the graph of the cumulative incidence curves
}
\description{
Create cumulative incidence plots
}
\examples{
Melanoma <- MASS::Melanoma
library(survival)
library(dplyr)
#Make the variable in the format for competing risks
Melanoma2 <- 
Melanoma \%>\% 
mutate(status = as.factor(recode(status, `2` = 0, `1` = 1, `3` = 2)))
#Make the status a labeled factor
Melanoma2 <- Melanoma2 \%>\% mutate(status = factor(status,levels = c("0","1","2"),
   labels = c("Censored","Melanoma","Other Causes")))
#Define the model
ajfit <- survfit(Surv(time, status) ~1, data = Melanoma2)
#Plot it
ci_graph(ajfit)

}
