% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_riskFinal.R
\name{ci_risk}
\alias{ci_risk}
\title{Create risk tables for competing risks}
\usage{
ci_risk(model_name, text_size = 4)
}
\arguments{
\item{model_name}{A competing risks object of the class survfitms survfit}

\item{text_size}{The size of the text in the risk table}
}
\value{
A GGPlot object with the risk table for competing risks
}
\description{
Create risk tables for competing risks
}
\examples{
Melanoma <- MASS::Melanoma
library(survival)
library(dplyr)
#Make the variable in the format for competing risks
Melanoma2 <- 
Melanoma \%>\% 
mutate(status = as.factor(recode(status, `2` = 0, `1` = 1, `3` = 2)))
#Make the status a labeled factor
Melanoma2 <- Melanoma2 \%>\% mutate(status = factor(status,levels = c("0","1","2"),
   labels = c("Censored","Melanoma","Other Causes")))
#Define the model
ajfit <- survfit(Surv(time, status) ~1, data = Melanoma2)
#Plot it
ci_risk(ajfit)

}
