% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_and_conf_int.R
\name{med_and_conf_int}
\alias{med_and_conf_int}
\title{Print text of the median survival value and its subsuquent 95\% confidence interval}
\usage{
med_and_conf_int(
  model_name,
  digits_all = 1,
  digits_med = NULL,
  digits_ci = NULL,
  unit_text = ""
)
}
\arguments{
\item{model_name}{the unstratified survival model}

\item{digits_all}{number of digits to round the median and CI too}

\item{digits_med}{number of digits to round mean to. Must specify digits_ci}

\item{digits_ci}{number of digits to round CI to. Must specify digits_med}

\item{unit_text}{unit for the median survival. Comes right after the median in the text.}
}
\value{
text with the median survival value and its subsuquent 95\% confidence interval
}
\description{
Print text of the median survival value and its subsuquent 95\% confidence interval
}
\examples{
fit <- survival::survfit(survival::Surv(time, status) ~ 1, data = survival::aml)
med_and_conf_int(fit, digits_med = 0, digits_ci = 1, unit_text = "years")
}
