% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_surv_flex.R
\name{stack_surv_flex}
\alias{stack_surv_flex}
\title{Stack tbl_surv_to_flex tables}
\usage{
stack_surv_flex(
  table_list,
  label_names = NULL,
  label_header_names = NULL,
  text_font_size = 9,
  table_title = "",
  bold_title = TRUE,
  title_font_size = 12,
  table_footnote = "",
  footnote_font_size = 9,
  missing_value_replace = "—",
  border_color = "#D3D3D3",
  header_alignment = "center"
)
}
\arguments{
\item{table_list}{A list of flextables created by tbl_surv_to_flex. Formatting in these tables will be lost.}

\item{label_names}{A named character vector the consists of the labels for the variables or outcomes in the "Characteristic" column.}

\item{label_header_names}{A named character vector the consists of the labels for the headers.}

\item{text_font_size}{A numeric that defines the size of the text in the table. Default is 9.}

\item{table_title}{A character that defines the title of the table. Default is empty.}

\item{bold_title}{A boolean that defines if the title of the table will be bold or not. Default is TRUE, which gives a bold title.}

\item{title_font_size}{A numeric that defines the size of the text in the title. Default is 12.}

\item{table_footnote}{A character that defines the footnote of the table. Default is empty.}

\item{footnote_font_size}{A numeric that defines the size of the text in the footnote Default is 9.}

\item{missing_value_replace}{A character that defines what values will replace missing values in the table. Example when the survival never gets below 50\%, the median survival is missing. Default is "—".}

\item{border_color}{A character the defines the the color of the borders in the table. Default is #D3D3D3.}

\item{header_alignment}{A character value that is either 'left', 'right', 'center', or 'justify' that defines the alignment of the headers.}
}
\value{
A flextable consisting of the stacked flextables
}
\description{
Stack tbl_surv_to_flex tables
}
\examples{
Melanoma <- MASS::Melanoma
library(survival)
library(dplyr)
Melanoma2 <- 
  Melanoma \%>\% 
  mutate(status = as.factor(recode(status, `2` = 0, `1` = 1, `3` = 2)))
#Make the status a labeled factor
Melanoma2 <- Melanoma2 \%>\% mutate(status = factor(status,levels = c("0","1","2"),
                                                  labels = c("Censored","Melanoma","Other Causes"))) \%>\%
  mutate(sex = factor(sex,levels = c("0","1"), labels = c("Female","Male")))

#Fit the models
ajfit <- survfit(Surv(time, status) ~1, data = Melanoma2)
ajfit_strata <- survfit(Surv(time, status) ~sex, data = Melanoma2)

#Create the tables for the specfic times
times2 <- gtsummary::tbl_survfit(ajfit, times = c(1000,2000,3000))
times2_strata  <- gtsummary::tbl_survfit(ajfit_strata, times = c(1000,2000,3000))

#Use the tbl_surv_to_flex function
xv <- ez.surv.viz::tbl_surv_to_flex(times2)
xv_strat <- ez.surv.viz::tbl_surv_to_flex(times2_strata)

#Combine in a list
comb_list <- list(xv,xv_strat)

#Stack them
stack_surv_flex(comb_list, label_names = c("Overall" = "All People"))

}
