% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_ci_graph_final.R
\name{strat_ci_graph}
\alias{strat_ci_graph}
\title{Create stratified cumulative incidence plots}
\usage{
strat_ci_graph(
  model_name,
  line_size = 1,
  outcome_colors = NULL,
  col_palette = "Dark 3",
  xlab_name = "Time",
  ylab_name = "Cumlative Incidence",
  conf_int = TRUE,
  conf_int_alpha = 0.2,
  legend_text_size = 10,
  legend_title_text_size = 12,
  legend_size = 0.5
)
}
\arguments{
\item{model_name}{A competing risks object of the class survfitms survfit}

\item{line_size}{A numeric containing the value of the line thickness. Default is 1.}

\item{outcome_colors}{A vector consisting of the color names/hexadecimals. Default is colors from Dark 3 palette}

\item{col_palette}{A string containing the name of the hcl palette to use. See hcl.pals() for names. Default is Dark 3}

\item{xlab_name}{Label for the x-axis. Default is "Time"}

\item{ylab_name}{Label for the y-axis. Default is "Cumulative Incidence"}

\item{conf_int}{Determines whether a confidence interval should be drawn for each curve. Default is TRUE.}

\item{conf_int_alpha}{A value between 0 and 1 that determines the alpha of the ribbon confidence interval}

\item{legend_text_size}{A number that determines the size of the text in the legend. Default is 10.}

\item{legend_title_text_size}{A number that determines the size of the title in the legend. Default is 12.}

\item{legend_size}{A number that determines the size of the legend in centimeters. Default is 0.5.}
}
\value{
A GGPlot object with the graph of the cumulative incidence curves
}
\description{
Create stratified cumulative incidence plots
}
\examples{
Melanoma <- MASS::Melanoma
library(survival)
library(dplyr)
Melanoma2 <- 
 Melanoma \%>\% 
 mutate(status = as.factor(recode(status, `2` = 0, `1` = 1, `3` = 2)))
#Make the status a labeled factor
Melanoma2 <- Melanoma2 \%>\% mutate(status = factor(status,levels = c("0","1","2"),
         labels = c("Censored","Melanoma","Other Causes"))) \%>\%
 mutate(sex = factor(sex,levels = c("0","1"), labels = c("Female","Male")))
#Define the model
ajfit <- survfit(Surv(time, status) ~sex, data = Melanoma2)
#Plot it
strat_ci_graph(ajfit)

}
