% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_graph.R
\name{ci_graph}
\alias{ci_graph}
\title{Create cumulative incidence plots}
\usage{
ci_graph(
  model_name,
  line_size = 1,
  outcome_colors = NULL,
  col_palette = "Dark 3",
  factor_order = NULL,
  xlab_name = "Time",
  ylab_name = "Cumlative Incidence",
  legend_name = "Outcome",
  conf_int = TRUE,
  conf_int_alpha = 0.2,
  legend_text_size = 10,
  legend_title_text_size = 12,
  legend_size = 0.5,
  legend_pos_x = 0.15,
  legend_pos_y = 0.8
)
}
\arguments{
\item{model_name}{A competing risks object of the class survfitms survfit}

\item{line_size}{A numeric containing the value of the line thickness. Default is 1.}

\item{outcome_colors}{A vector consisting of the color names/hexadecimals. Default is colors from Dark 3 palette}

\item{col_palette}{A string containing the name of the hcl palette to use. See hcl.pals() for names. Default is Dark 3}

\item{factor_order}{A vector of the outcomes in the order that the user wants them to be listed in the legend. First vector will be at the top, second will be after it and so on and so forth. Please be careful and only use with a labeled factor of outcomes.}

\item{xlab_name}{Label for the x-axis. Default is "Time"}

\item{ylab_name}{Label for the y-axis. Default is "Cumulative Incidence"}

\item{legend_name}{Lbael for the legend. Defauly is "Outcome"}

\item{conf_int}{Determines whether a confidence interval should be drawn for each curve. Default is TRUE.}

\item{conf_int_alpha}{A value between 0 and 1 that determines the alpha of the ribbon confidence interval}

\item{legend_text_size}{A number that determines the size of the text in the legend. Default is 10.}

\item{legend_title_text_size}{A number that determines the size of the title in the legend. Default is 12.}

\item{legend_size}{A number that determines the size of the legend in centimeters. Default is 0.5.}

\item{legend_pos_x}{A number the determines the x position of the legend inside the graph. Default is 0.15}

\item{legend_pos_y}{A number the determines the y position of the legend inside the graph. Default is 0.8}
}
\value{
A GGPlot object with the graph of the cumulative incidence curves
}
\description{
Create cumulative incidence plots
}
\examples{
Melanoma <- MASS::Melanoma
library(survival)
library(dplyr)
#Make the variable in the format for competing risks
Melanoma2 <-
Melanoma \%>\%
mutate(status = as.factor(recode(status, `2` = 0, `1` = 1, `3` = 2)))
#Make the status a labeled factor
Melanoma2 <- Melanoma2 \%>\% mutate(status = factor(status,levels = c("0","1","2"),
   labels = c("Censored","Melanoma","Other Causes")))
#Define the model
ajfit <- survfit(Surv(time, status) ~1, data = Melanoma2)
#Plot it
ci_graph(ajfit)

}
